MissileEx
MissileEx is a new projectile trying to expand upon Missile with as little drastic modification to it as possible (still very small changes here and there will probably be needed tho). Currently, it adds RandomHLaunchAngleStart and RandomHLaunchAngleEnd. Since this is an extension upon Missile, everything else in Missile is normally doable.

The current missile facing minus Start is the starting point and the missile facing plus End is the ending point, then it picks an angle at random from this range.