Fixed some Crashes + TransportProperties Changes
My previous changes to TransportProperties caused crashes, for a reason very obvious to me now. Not only is this crash fixed, but proper script handling of multiple cargoholds on a single actor should now be implemented and working.

Not sure why I was getting a crash with techTree being null, I imagine it's due to some of my other changes to ProductionQueue, but this is now also fixed.